Add files via upload
Blossom Boutique - Flower Shop Website

This is a simple, beautiful flower shop website built with HTML, CSS, and JavaScript. It allows users to browse a selection of flowers, add them to their shopping cart, and proceed to checkout. Additionally, the website includes a Contact Form for users to reach out and an Admin Panel for managing the products.